requirements/base.txt
coreapi==2.3.3
django==4.2.3
django-cors-headers==4.2.0
django-environ==0.10.0
djangorestframework==3.14.0
django-filter==23.2
djangorestframework-simplejwt==5.2.2
django-jazzmin==2.6.0
drf-yasg==1.21.7
psycopg2==2.9.6
